Update to V2.0

+++ OT_Emulation_Data_Broker/DataBroker/Sim_Control.c
@@ -1,5 +1,6 @@
 #include "Sim_Control.h"
 #include "Sem_Stop.h"
+#include "init_Server.h"
 
 void User_Control(void){
     
@@ -17,39 +18,51 @@ void User_Control(void){
         if (invar == 'x' || invar == 'X'){
             printf("Stopping Simulator\n");
             sem_post(stop);
+            sleep(1); /*slow down the kill to let threads close out safely*/
             break;
         }
 
     }
 }
 
-
+/* Notes: Its important not to fork the process if we are not executing the simulator from the DB.
+The relationship between the parent process and child has to change depending on external or internal
+execution of the simulator. This is the simplest and most stable solution. */
 void *Sim_Control(){
-    
+
     pid_t pid;
     char *args[2];
-    args[0] = "ans_runtime_15mar21";
+    args[0] = SimName();
     args[1] = NULL;
     
-    /* fork process */
-    pid = fork();
-    
-
-    switch(pid){
-        case -1:
-            /* Fork failed*/
-            perror("Fork failed");
-            break;
-        case 0:
-            execv("ans_runtime_15mar21",args);
-            printf("Simulator has failed to load! \n");
-            break;
-        default:
-            printf("Simulator running\n");
-            User_Control();
-            break;
+    /* checking if Simulink external simulator was selected*/
+    if(!strcmp(args[0],"Simulink")){
+        printf("External simulator selected. \n****You may now start the simulator****\n");
+        User_Control();
     }
-    kill(pid,SIGTERM);
+    else{
+        /* fork process */
+        pid = fork();
+
+        switch(pid){
+            case -1:
+                /* Fork failed*/
+                perror("Fork failed");
+                break;
+            case 0:
+                /* Child process will run the user control */
+                printf("Starting Simulator\n");
+                execv(args[0],args);
+                printf("Simulator has failed to load! \n");
+                break;
+            default:
+                /* Parent starts the user control */
+                printf("User Control Initializing\n");
+                User_Control();
+                break;
+        }
+        kill(pid,SIGTERM);
+    }
 
 pthread_exit((void *)0);
 }
